extractor/pyproject.toml
[project]
name = "extractor"
version = "0.1.0"
description = ""
authors = [{ name = "sigalor", email = "nalenz@protonmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "langextract (>=1.0.9,<2.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.32.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
]

[tool.hatch.build.targets.sdist]
include = ["src/extractor"]

[tool.hatch.build.targets.wheel]
include = ["src/extractor"]

[tool.hatch.build.targets.wheel.sources]
"src/extractor" = "extractor"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
